project_finance_backend/pyproject.toml
[tool.poetry]
name = "project-finance"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Rofar02 <lexa95k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.127.0"}
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.45"
black = "^25.12.0"
alembic = "^1.17.2"
asyncpg = "^0.31.0"
passlib = "^1.7.4"
bcrypt = "^5.0.0"
pyjwt = "^2.10.1"
email-validator = "^2.3.0"
aiogram = "^3.13.0"
cryptography = "^42.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
